Feature/PG-7 interafaccia barra di stato e queue-logger (#42)
* afeat(score):ggiiunto score, e display delle varie infomazioni

* feat(logview):added the log sub view

* feat(subViewLogger): aggiunto la finestra per il logger

* fix(convert string):fissato il waring delle stringhe

* fix(tests): tolti gli errori dai tests

* fix(pr): adeguato il codice hai commenti

+++ tests/binary-tree.cpp
@@ -16,7 +16,7 @@ namespace test {
         }
 
         int *array = h.asArray();
-        for (int i = 0; i < h.getSize(); i++) {
+        for (size_t i = 0; i < h.getSize(); i++) {
             assert(array[i] == (int) i);
         }
         assert(h.max() == 99);
