+++ tests/vector.cpp
@@ -34,7 +34,7 @@ namespace test {
 
     void correct_size() {
         Vector<int> v;
-        for (int i = 0; i < 20; i++) {
+        for (size_t i = 0; i < 20; i++) {
             v.push_back(i);
             assert(v.size() == i + 1);
         }
